refactor(datamapper): split field context menu hooks and tests
Fixes #3165
