+++ packages/ui/src/components/Document/actions/FieldContextMenu/types.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react';
+
+import { MenuGroup } from '../FieldContextMenu';
+
+export interface MenuContributor {
+  groups: MenuGroup[];
+  modals: ReactNode;
+}
